Check PSNE on corefx deserialization, check round-trip on netfx
This extends @hughbe's helper from #20984 for testing that exceptions
throw PSNE on deserialization, so that for IsFullFramework rather than
abort the test, it tests round-tripping.

The test that PR included is changed to use this adaptive test, and
some round-trip tests that @morganbr removed in #19742 are brought
back using this new helper.