Remove concept of bogus from types and most symbols.
Since bogus types can't get as far as the dynamic binder, the only case
where bogus can be true is when a non-indexer parametrized property,
defined in a language where that is allowed, is encountered.

In other cases computing bogosity involves recursive checks that always
hit false.

Give PropertySymbol a separate Bogus property to use where it could be
set, and remove all other uses of GetBogus and similar.

Entails removal of ERR_BogusType, contributes to #22470