Remove useMethInstead from property symbols and consider always true.
useMethInstead should be set in AddPropertyToSymbolTable when bogus is
set, but is not.

However, the only time it is looked for the only possible value is true
as that is the only condition under which a symbol can be bogus, so
remove it, and the path for it being false.

This now means that ErrorCode.ERR_BindToBogusProp1 or
ErrorCode.ERR_BindToBogusProp2 will be used when appropriate.

Fixes #22534