Make OpenSSL verify self-issued signature integrity to match Windows
Windows checks signatures on self-issued certificates (to prove they are
self-signed), but OpenSSL doesn't by default, opting to save on the CPU.

Before this change the behavior was:

* Windows: NotSignatureValid
* macOS: UntrustedRoot
* Linux: UntrustedRoot

With this change it becomes

* Windows: NotSignatureValid
* macOS: UntrustedRoot
* Linux: NotSignatureValid | UntrustedRoot

While additional work could be done to not report UntrustedRoot for Linux
chains, it doesn't seem as worthwhile.

macOS will continue not reporting NotSignatureValid, because there's no API
to allow changing the behavior.