Make Expr.Type in Microsoft.CSharp non-virtual.
Since those types that compute this do so in a limited number of places
and since only a small number never set it, the advantage of having the
member virtual is low.

Conversely, since the property is accessed a very large number of times
the advantage of having the getter a simple (hidden) field access which
would be fast and more easily inlined, is greater.