Refactor GetTypeForErrorMetaObject
We only need to extract the first item in args in one case, so pass the
array rather than that first element.

Replace as used just as test with is.

In action is CSharpInvokeConstructorBinder branch, first argument is
guaranteed to be a Type, or there would have been an ArgumentException
earlier, so no need to guard against it being otherwise.