Fix perf regressions in BigInteger
We incurred some perf regressions in BigInteger(byte[]) and ToByteArray() due to the span changes.  This mostly fixes them.  In fact, BigInteger(byte[]) is now a bit faster than it was before the span changes.  There is still a small regression in ToByteArray() for values <= int.MaxValue, but for values larger than that, there's a similarly-sized improvement over the original version; since BigInteger is all about, well, big integers, this looks like the right trade-off, and it allows us to continue avoiding to triplicate the same code path for ToByteArray/GetByteCount/TryWriteBytes.