Expand Deflate64 WindowSize
The WindowSize being used in the old managed deflate implementation that I modified to decompress Deflate64 is too small. Deflate64 can have lengths up to 64kb as well as distances up to 64kb when doing lookback. In my initial implementation I disregarded the increased length, allowing potentially 64kb of data to be written to an outputwindow that only has as little as 258 bytes of open space. This causes the previously inflated yet-to-be-flushed data to be overwritten, causing invalid output.

The fix is simple: increase the WindowSize to fit the new length/distance maximums. Becuase of the way our managed Deflate implementation is written, it's not possible to simply flush the OutputWindow after already determining the lookback without a large rewrite, so I opted for the increased WindowSize solution instead.

Testing is unfortunately not simple because it's rare that a file will need such a high length/distance and we do not have a Deflate64 deflater so we have to rely on other programs for generating test input. I've got a file that allows reproduction of the failure, but it's 3gb and doesn't experience data loss until a little under the 1gb mark. Because of the large size of the test file, I've left it out.

I also took the chance to cleanup the s_extraLengthBits field. Currently, we attempt to read extra bits from input when given an invalid symbol for determining length. By shortening the arrays to only valid values, we instead throw an InvalidDataException early.