Prevent unnecessary AIA fetches of trusted roots
The "Simplify X509Chain" change no longer used userRootCerts or
systemRootCerts in FindCandidates. Since the eventual chain build would get
them from the trusted path there's not a functional difference, but
FindCandidates was doing an AIA fetch for the root certificates so it could
complete it locally.

So just keep those certs in the scan list to revert the network use regression.