Update TestService to log events
The managed TestService now has a simple ServiceBase TestService
that logs all the events it recieves to a file.

The project also as a simple TestServiceInstaller that will allow
for programatic install and removal of a service.

The main entry point provides a simple way to create/remove/run
the TestService.