Add ServiceBase tests
Update test project to use the new managed TestService instead of
the old native test service and adjust the existing tests to account
for those changes.

Add basic ServiceBase tests. It isn't possible to test all the different
events so only test the ones that we can reliably tests.

The communication between the service and the tests is via a text file
which isn't ideal but it is the simpliest option of the various options
to work. All the tests need to call Stop to ensure the log file is flushed
to disk and we have all the content we need in it.