Add Read/Write(Span) overrides to BufferedStream (#22948)
Copy existing Read/Write(array) implementations and just tweak them to work for span.  I purposefully kept the same logic as was there.  I also didn't just make the array implementations work in terms of the span ones so that the array ones still result in calls to the underlying stream's array-based methods.