Make ManagedTests.DynamicCSharp…regclass029 test assert finalization (#22914)
#5677 fixed a race in this test that was causing it to sometimes fail
by finalizing too soon, but also moved the assertion of finalization
into the finalizer, so it would only assert that the finalizer had run
if the finalizer had run.

Move the assertion back to the point where the finalizer should be know
to have run, but not in the finalizer itself. (That PR still fixed the
actual race that stopped success being detected).