Refactor Microsoft.CSharp's error handling. (#22748)
* Use ctor in CError instead of Initialize method.

* Remove CErrorFactory

No state or alternative implementation to give factory pattern an advantage.

* Replace CController SubmitError with instance CError.Submit method.

Has same effect as sole implementation of SubmitError.

* Remove CController and RuntimeBinderController

No longer doing anything.

* Remove CError

Just creates a string for RuntimeBinderException, so create that string directly.

* Remove IErrorSink

Looser coupling offered is not taken advantage of. TIghter coupling is simpler, just have ErrorHandling do it.

* Simplify SymbolLoader ctor.

More self-sufficient; doesn't need parameters.

* Delete LangCompiler

Only functionality is initialising SymbolLoader; can be done in base class.

* Remove CSemanticChecker.GetSymbolLoader

Just call the SymbolLoader property it calls into.

* Replace CSemanticChecker.GetErrorContext with call to property it calls into.

* Remove ErrArgRef

Differs from ErrArg in requesting that error location be included, but dynamic binder can never do that.

* Remove AddRelatedSymLoc and AddRelatedTypeLoc

Both are no-ops

* Remove _swtWrongCount and stop tracking such symbols.

Never made use of.

* Make Error() only public method on ErrorHandling

No longer anything attempted with CParameterizedError between creation and submission.

* Simplify ErrorHandling.Error

Remove CParameterisedError which is just a wrapper on the code and array of arguments passed, and create the error to throw immediately.

* Localise knowledge of UserStringBuilder to ErrorHandling.

* Have ErrorHandling.Error return rather than throw an exception.

Move the throw up to the caller, making it clear that no code can
follow it.

In the case of a method that would throw for every path, similarly have
it return the exception, so it gets thrown at its caller in turn.

* Remove CreateUserLogOpError

No callsites.

* Remove MemberGroup.SetInaccessibleBit()

No callsites

* Remove MemberLookup.SwtInaccessible

No call sites.

* Remove canConvertInstanceParamForExtension

No callsites.

* Remove path on VerifyNamedArgumentsAfterFixed returning false

As it always returns true.

* Have VerifyNamedArgumentsAfterFixed return value instead of using out parameter

* Remove BindingContext.ReportUnsafeErrors

Only ever set to false in a path within ExpressionBinder.checkUnsafe that will always throw first if it is true, and hence never false.

* Remove ExpressionBinder.ReportTypeArgsNotAllowedError

Move into only caller, which has other code that is similar and gains little from the spit.

* TryReportLvalueFailure as void

Can only return false, so return type is meaningless.

* Remove ErrorCodes from secondary errors

These codes would only be used after another error, which for the
dynamic binder means not at all, since the error immediately throws.

Now they are unreachable, so remove. Contributes to #22470