Redefine GetCTypeArrayFromTypes in terms of Type[]
Never called with null, so replace the null path with an assertion.

All but one call has type IList<Type> parameter as Type[], and that
call is improved by changing it from List<Type> to type[]. Do so, and
change GetCTypeArrayFromTypes to take an array, returning Array.Empty
when appropriate (a common case) and using array iteration for the
copy, otherwise.