Remove code in Microsoft.CSharp related to extension and external methods  (#22999)
* Remove paths for extension methods.

Dynamic doesn't support extension methods, except for static-style
calls, so isExtension is never true.

Remove it and all paths that depend on it being true.

* Remove error conditions relating to extension methods.

Demonstrated unreachable. Contributes to #22470

* Remove MethodSymbol.isExternal

Never set to true. Remove it and all paths depending on it.

* Remove check for MemLookFlags.ExtensionCall flag

Never set, so will never pass. Involves removing
ERR_NoSuchMemberOrExtension, contributes to #22470