Define AddMethodToSymbolTable as taking MethodBase
AddMethodToSymbolTable is called with a MemberInfo that can be either a
MethodInfo or a ConstructorInfo, and takes different paths for each.

Much of what these paths do is a duplication of calls that are
available through those types common MethodBase base.

Change the parameter type to MethodBase and use it as much as possible.