Ensure packages continue to build for NETCoreApp2.0
These are for packages that are outside of the shared framework.  We shouldn't make them
cross-compile for netcoreapp2.1 just because master has moved to that.  Instead we'll make
them continue to target netcoreapp2.0.

Two cases we want to cross-compile are OpenSsl and Cng since these have already started
taking advantage of API added in NCA2.1 (span).  For this we'll cross-compile and redist the
old netcoreapp2.0 implementation.

Two other special cases are System.Memory and System.Runtime.CompilerServices.Unsafe.
In these cases we wanted to keep them targeting netcoreapp (netcoreapp2.1) and
make their portable implementations apply for netcoreapp2.0.