Disable two collections tests on desktop (#23200)
These tests are verifying that `{Equality}Comparer<T>.Default` is idempotent.  But whereas on core they are now initialized with beforefieldinit, on desktop they're lazily initialized with optimistic concurrency in a manner that can result in two threads each seeing their own distinct instance.  That invalidates these tests, which spordically fail in CI.