Turn on setting GIT URL in File Version
This turns on tagging each CoreFX DLL built GitHub URL
of the source code that was used to build it.

You now get something that looks like the folloiwng when you do a filever -v

ProductVersion  4.6.25612.0 @BuiltBy: vancem-VANCEM4 @SrcCode: https://github.com/dotnet/corefx/tree/a329e252594f375348a8f06a3675606f5c03ce3f

Note that most of the work for this is in the buildtools, this simlpy sets the GitHubRepositoryName variable that turns it on.