Fix handling of flock in FileStream on Unix (#23235)
On Unix, we implement FileShare in FileStream with flock.  On some file systems, flock returns EACCES, which causes us to fail to allow usage of FileStream at all.  This commit changes the code to only handle the one error we actually care about, EWOULDBLOCK, which indicates that the operation would block and hence that sharing failed.  All other errors can be ignored, as this is only advisory / best-effort.