Avoid TypeLoadException in async code
Factor out the "eat preamble" code into its own method (that can be
reused by both the sync and async code) to prevent the
`ReadOnlySpan<byte>` preamble from being captured in a field of the
async state machine. This avoids the System.TypeLoadException : A value
type containing a by-ref instance field, such as Span<T>, cannot be used
as the type for a class instance field.