Remove code in Microsoft.CSharp for casting from method groups.
As a method group cannot be cast to object, the casting from method
group to delegate that can be done in static code is not relevant, and
code related to such casts is never reached.

Remove it. Includes removal of ExprFuncPtr, and ERR_MethGrpToNonDel,
ERR_BadRetType, ERR_DelegateOnNullable & ERR_BadDelegateConstructor,
contributes to #22470