Avoid char[] allocation in GetComputerName
Avoid the `char[]` allocation in the `GetComputerName` Win32 interop
code by using `stackalloc`, which is used to implement
`Environment.MachineName` on Windows.

While we're at it, the size of the buffer can be reduced. The current
length is 256, but this is only for historical reasons when the buffer
needed to be large enough to support Unix host names when it'd be
calling a PAL implementation in CoreCLR on Unix. We're no longer calling
a PAL implementation, so we can reduce the size of the buffer to what's
required for this function. According to the MSDN documentation for
`GetComputerName`, "the buffer size should be large enough to contain
MAX_COMPUTERNAME_LENGTH + 1 characters."

Also, update `GetComputerName` to return null on error, so the calling
code in `Environment.Windows.cs` will throw `InvalidOperationException()`
as documented.