Fix output buffer assumptions in CSP symmetric transforms.
The CAPI DecryptData and EncryptData methods made assumptions about the
calling patterns, in particular that the output buffer would be perfectly sized for
the desired output.

This change adds tests which violate those assumptions, but work on netfx and
with the corefx CNG implementation. Then follows up on making the tests pass.

NetFX idiosyncrasies:
AesCryptoServiceProvider.CreateDecryptor on netfx guards against the key not
being specified, but no other Aes class does.

TripleDESCryptoServiceProvider on netfx doesn't check if the output buffer is
too small, so the ArgumentOutOfRangeException ends up being a
CryptographicException instead.