Fix handling of last number in IPv6 address
A previous fix for mishandling of certain IPv6 addresses ended up introducing another bug, where we may fail to fail parsing on an address that has too many numbers in its last sequence.  The issue is that, if the string doesn't end with ']', we're not looping around again in the parsing loop, where the beginning of the loop would see the sequence thus far as being too long and error out (the previous fix included removing the adding of a ']' if it didn't exist).  The fix is just to do the check again at the end.