Fix behavior of Dns.GetHost* with empty string
On Windows, passing an empty string to gethostbyname/GetAddrInfoW results in the host's information being used; that behavior then bubble's up through (and is documented for) the Dns.GetHost* APIs in .NET.  On Unix, the native functions being used don't provide that same behavior, instead returning an error.

This just special-cases empty string in the Unix implementation.