Re-enable retrying checkout (#23801)
Patching the Windows OS seems to have resolved the checkout hangs.  It appeared to have been hanging in some Windows filesystem code.  It also may have been because of git performance issues in 2.11+ when fetching large numbers of refs, which was resolved by a recent change to only fetch the ref relevant to the current PR.