Xml schema dependency removal for AOT compilers (#23867)
This changes fixes a size issue in AOT compiled code which uses XmlReader. Using XmlReader for just reading XML without validation still pulls in all of the XmlSchema code. This makes the AOT compilation slow (it's more than 400KB of IL) and the resulting app is larger (.NET Native X86 measures the difference at more than 800KB of native code).
The change relies on an assumption, that apps which don't need validation will not call the XmlReaderSettings.ValidationType setter. As such it can be reduced away and the change then ties all the dependencies needed for validation to that setter. So if the setter itself is not called XmlSchema is not pulled in.