Remove dead path from BigInteger ctor
If a BigInteger is created from a byte array of more than 4 bytes, and
all of them are zero, then a path is taken assign the zero constant to
the constructed struct.

However, as trailing zero bytes are skipped, no all-zero sequence will
be treated as being larger than zero bytes, so the path can never be
hit.

Remove it.