Fix WinHttpHandler when building certificate chain
Unlike SslStream, WinHttpHandler was not using all the certificates
obtained from the TLS/SSL handshake when building the chain. In
addition to the end-entity server certificate, there were other certs
available in the native cert handle that could be passed along
to the chain build operation.

As part of fixing this, I refactored the Common schannel and crypto
interop code so that it can be used in both SslStream and
WinHttpHandler.

I tested this manually because this cannot be easily tested in the CI
system. Added traces in WinHttpHandler to support this.

Fixes #23827
Fixes #4473