Fix globalization tests to work with setting customization (#24052)
We used new CultureInfo in many places to check the NumberFormatInfo properties values . these tests can fail if the user customized the  user settings on the running machine. The fix here is to use GetCultureInfo which always ignore the change in the settings and return the original values we need to test against.