Fix RHEL6 linux/in.h file detection (#24100)
On RHEL6, the build got broken by my recent fix of the in_pktinfo /
ip_mreqn detection fix. The reason is that on that platform, linux/in.h
cannot be included on its own, the sys/socket.h needs to be included
before that to provide definitions of some structures. In fact, Linux
doc states that the sys/socket.h should be included before the
linux/in.h. So I am fixing that by adding that include to the detection
code.