Avoid blocking ThreadPool threads in cancelable NamedPipe*Stream.ReadAsync on Unix
Named pipes on Unix are implemented via sockets.  When a ReadAsync is issued, that it turns results in a ReceiveAsync on the socket.  But today we lack a ReceiveAsync overload that takes a CancellationToken, and until such an overload exists, we need a workaround.

Today's workaround involves polling and blocking a thread pool thread doing a synchronous read; that's very expensive, especially when many such calls are made across one or more pipes and we end up blocking many threads.

This PR provides an alternative workaround.  Instead of queueing a blocking poll to the thread pool, we issue a zero-length async receive so that we're notified when data is available, and we concurrently monitor the cancellation token.  This is more efficient when cancellation doesn't occur.  If cancellation does occur, we potentially "leak" the read request until data is available (or the socket closes), but cancellation is relatively rare, and even more rare is reusing the stream after cancellation has occurred, so this shouldn't be much more than a small fixed overhead.