Fix offical Alpine build - sync arguments
It turns out that the sync.sh command needs to be passed the -SkipTests
option too so that it doesn't try to sync core-setup packages that
are used only for running tests. Since there are no core-setup packages
for Alpine yet, the sync was failing.
I have also noticed that there was a recent change in the sync arguments
that I have not discovered after rebase of my changes to master. So I
have added the new option there.