Fix DebuggerDisplay for XmlNode
Somewhere along the way XmlNode's debuggerDisplayProxy property was removed.  That doesn't break the build, as it's only used via reflection, but it does break the debugger display tooltip.

This just copy-and-pastes the corresponding implementation from reference source to fix it.  The only other change I made was renaming "node" to "_node" and making it readonly.