[UWP]allow all security flags on named pipe client
Perhaps the more proper way to deal with this is to define proper
constants and enumerations but that needs to be done for the whole IO
area/usage. However, even after that likely we should change the signature of the
method since there will be some values that are not supported with the
current signature.

Opportunistic fix: RunAsClient was not raising proper exceptions in case
of error.