Revert bad PR that was merged to soon
See #24064. The checked in fix is wrong because it changes the behavior when .NET Native is being used such that a nonsensical exception will be thrown which would never be thrown when .NET Native is not being used. These kinds of errors are really hard for developers to debug.

Instead we will let .NET Native fail in the normal way when rd.xml data is not available. This will direct people to add metadata appropriately when they need to get this to work.

@AtsushiKan @danmosemsft Given the limitations in .NET Native, should we avoid adding code that uses Reflection in new .NET Core code going forward? (Btw, this is truly a case where the type is completely unknown to the framework and could be almost anything.)

If we are going to continue to use Reflection in new .NET Core code, then should it be the case that:

The runtime gives enough details to the application so the developer can correctly figure out what to add to rd.xml to get things working?
If the app developer doesn't do this, then it will continue to not work
Code that uses Reflection should provide helpful message as to what is not supported on .NET Native?
That is, it won't just fail in some non-obvious way like in this PR.