Remove dead code concernin TypeParameterType/Symbol from Microsoft.CSharp (#23278)
* Remove cases concerning TypeParameterType conversions

Only relevant when compiling generic methods, not in dynamic code.

Entails removal of ERR_TypeVarCantBeNull. Contributes to #22470

* Remove case for TypeParameterType in AllPossibleInterfaces

Since both callsites ensure the type is a struct, class or interface type, it can only be called with an AggregateType or NullableType

* Remove _bHasRefBound, _bHasValBound and _pDeducedBaseClass

Only ever false, false and null respectively.

* Remove base and boxing type parameter conversions.

Only relevant when compiling generic methods, so never hit in dynamic code.

* Don't consider type parameter types in GetRefEqualSigs

Only comes up when statically compiling generic methods.

* Remove unnecessary gotos

* Dont consider type parameter types in HasImplicitReferenceConversion

Only comes up when static compiling generic methods.

* Reduce repeated checks within HasImplicitReferenceConversion

(Some of these will become better still when some methods that only return true for AggregateType are made sealed and virtual within it)

* Remove unnecessary identity check form HasImplicitReferenceConversion.

Never true when it is called, and if it was true the method should return true anyway.

* Move check that pDest is AggregateType

As all true-returning branches depend on it in some way, and it simplifies the rest.

* Replace if ladder on types of AggregateType into switch on symbol's AggKind

* Remove TypeParameterType case from MemberLookup.Lookup

Cannot happen outside of static compilation of generic method.

* Remove type parameter case from SearchSingleType

Cannot happen in dynamic code.

* Remove TypeVarsAllowed and ExtensionCall from MemLookFlags

Unused

* Remove GetInterfaceBounds from TypeParameterType and TypeParameterSymbol

Always null and no longer used.

* Remove GetEffectiveBaseClass from TypeParameterType/Symbol

Always null, and unused. (Removing either still results in null being
the value that would be found, or removing a case that would throw if
it was hittable).

* More consistent spacing in switch (respond to PR feedback).