Removed obsolete System.Net.Http NETFX related files
Due to PR #23611, the master branch no longer builds a NETFX binary for
System.Net.Http. Instead this is now handled only in the release/2.0.0
branch as part of the "NETStandard support package".

This effectively means that we have already stopped OOB'ing
System.Net.Http on .NET Framework platforms as of .NET 4.7.1 since that
version of .NET Framework supports all NETStandard2.0 APIs.  And .NET
Framework versions prior to that (4.6.x, etc.) are handled with the OOB
System.Net.Http.dll that is part of the NETStandard support package.