Port missing System.Net.Http NETFX fixes
Issue #23702 showed some Http test failures on NETFX. This was because
PRs, #21325 and #21373 are missing from the release/2.0.0 branch.

These fixes are needed for the OOB System.Net.Http.dll that is being
built for the NETStandard Support Package.  This package is used when
NETStandard is targeted and the platform is NETFX but less than .NET
4.7.1.