We do not support Process.Start on a directory on Unix unless UseShellExecute is set. When UseShellExecute is set, we can call xdg-open to open a window showing the directory.
When Process.Start is passed the directory (via ProcessStartInfo.FileName) we will attempt to resolve it to a fully qualified path.
We should be testing the actual path we are going to give to execve.
To get the actual path we first check if the path is exact, then we try path relative to executable and then current directory.
Lastly we check in all directories provided in PATH to find ProcessStartInfo.FileName.

After finding this path, we were inadvertently still checking the input path before this conversion was done.
The fix is to make the check after converting the path.

fixes #24402