Update how we do filtering to avoid path comparsions
In some cases we end up copying the incorrect set of files because
the path comparisions fail. It appears that in some cases msbuild
will normalize paths for *.dll wildcard expansion and switch slashes
around which causes the path comparsion to fail. We should avoid
comparing paths at all so instead do the filtering based on filename.