Update places to compare file names and not paths
There is an msbuild issue https://github.com/Microsoft/msbuild/issues/2620
that causes some paths to normalized where they weren't before and
this normalization causes conditions to start changing and excludes to fail.

In general it is not a good idea to depend on path comparision as a string
anyway so switched to using the file name in a few places instead of path.