Fix HttpWebRequest when sending with HTTP/1.0 (#24644)
* Fix HttpWebRequest when sending with HTTP/1.0

The HttpWebRequest.ProtocolVersion property was implemented. But it was
not checked when actually creating the request (which uses HttpClient
underneath). So, all requests were sent with HTTP/1.1.

This fix now uses the ProtocolVersion property and sends either
HTTP/1.0 or HTTP/1.1.

Note: HTTP/2.0 is currently not supported when setting the
ProtocolVersion property. An ArgumentException is thrown. This is the
existing .NET Framework behavior which was ported to .NET Core. We could
investigate changing this in the future to allow ProtocolVersion to be
set to 2.0.

Fixes #24597

* Skip test on UAP