Remove New APIs in Syndication. (#24814)
The PR removed all APIs that are added in Syndication on Net Core. We'd like to review those APIs thoroughly before exposing them.

Some public abstract async methods were changed to virtual because it's not allowed to have public abstract methods in the implementation assemblies while not exposing them in the contract assembly.