Update code in AwaitUnsafeOnCompleted to use interface matching
Given that the jit can now avoid boxing on some interface calls to value types,
generalize the patterns introduced in AwaitUnsafeOnCompleted in #14718 by
using interfaces instead of checking for specific types.

Also move the catch-all processing back in line as the jit can now fold the
interface tests early and not pull in the EH portion of the method unless it is
needed.

Signed-off-by: dotnet-bot-corefx-mirror <dotnet-bot@microsoft.com>