Add System.Management (#24719)
* First pass building

* Remove Instrumentation and WMIGenerator

* Resources to as used on CoreFX

Opportunistic: some small changes related to warnings

* Removed hard coded path to load wminet_utils

Also moved so DllImport to use Common version of files.

* Added ref, sln, and single basic test

Code builds and pass single test

* Add CoreFX source header

* Removing unused using directives

* clean src proj references

* A bit more test coverage (19%)

* Sanitizer pass with default settings

* Adding WMIGenerator

This allows methods to generate CodeDom for strong type access to WMI
objects.

* Missing files from previous commit

* Some extra clean-up missed by SSS

* Removing #ifdef dead code

* Add pkg project

* Removing more dead code

* Removing dead code and small improvement to tests

* Suppress CA1821 when generating PNSE assembly

The finalizers are being generated empty for PNSE assembly. In such
cases disable CA1821.

* Source changes from CR (so far)

* Targets and package issues

* Go back to project targets and settings

Per conversation w/ @weshaggard it should follow the same model as
System.DirectoryServices not the packages of dlls that do not exist on
desktop. This still is not expected to be fully correct but it puts the
projects closer to the targets that they will have in the end.

* Description for the System.Management package.

* Match supported pkg frameworks

* Skip tests on Windows Nano

* Missed files from previous commit

* Untabify files and a bit more tests

* Generate PNS Assembly for uap

* Add references needed by uap for PNS assembly

* PNS assembly message

* Fixing package issue