Add -g flag to generate debuginfo on unix platforms (#24979)
The windows build already includes /Zi /Zl as part of commit 920fd2f3
(PR #7840). It looks like it was simply missed on Unix.

This change also makes the native debug information closer to what
CoreCLR does on all platforms. See
https://github.com/dotnet/coreclr/pull/3445 for more information.

This is also needed for the end-to-end debuginfo generation as part of
source-build. See https://github.com/dotnet/source-build/issues/267