Add a simple way to test wmi remote (#24989)
Simply using an environment variable, WmiTestTargetMachine, to specify
the target machine. When running the tests ensure that the account
running the tests has rights on the target machine.

There is not much value in adding similar scope to other tests since
regarding remote connectivity they actually end up hitting the same
code, so keep this just for some key tests.