HashCode based on xxHash32 (#14354)
Works by maintaining the xxHash32 state variables (v1 -> v4, length) as well as a queue of values that fall outside of the block size (16 bytes/4 ints). Seed is assumed to be zero as there is no seed ctor in the proposal - however, it is possible to add a seed without changing the structure.

Tests against known xxHash32 vectors are provided ("abcd1234efgh5678...").