Remove checks for locals as possible source of lvalue errors.
Demonstrable two ways:

Firstly it is impossible to represent the sort of expression that cause
CS1604 or CS1605 in dynamic code.

Secondly as ExprLocal have Flags set to EXPRFLAG.EXF_LVALUE and that
flag never removed, the paths for expr.isLvalue within both checkLvalue
and TryReportLvalueFailure (before it loops) will mean an ExprLocal is
never examined within TryReportLvalueFailure, and so ReportLocalError
can never be called.

Entails removal of ERR_RefReadonlyLocal and ERR_AssgReadonlyLocal,
contributes to #22470