Switch to new PB_SkipTestBuild property to disable the test build
We need to always turn off tests for for some legs like alpine so
but setting the PB_SkipTests property in pipeline file doesn't
seem to correctly override the global value of this property. Perhaps
that is a bug in pipebuild but for now just using a different property
name to make forward progress.