Don't allocate in SubstTypeArray unless necessary.
As long as SubstTypeArray attempts a substitution, it will always
allocate and return a new TypeArray even if the contents are identical
to the source.

This in turn will cause calling substitutions to substitute an
identical CType.

Change so as not to allocate unless at least one substitution makes a
change.