Remove redundant check from flaky test (#25126)
That check for expected exception is already present in other tests but
it is very flaky in this specific test. The point of this test is about
method invocations and as long as the target process was terminated that
is verified so I'm simply removing the problematic step of the test.