MS.CSharp: Avoid allocation in AllocParams(int, TypeArray, int)
This is called to split the type arguments for nested classes between
those for the type itself and those for the outer type.

The most common case will return the empty array or the source array
(only really splitting for a type both within a generic type, that also
adds further type parameters), but getting there will allocate an array
and possibly do a lookup.

Catch and short-circuit on these cases.