Fix handling of unary operations on nullable enums in Microsoft.CSharp
The code that covered this case would put a NullableType where an
AggregateType would then be expected, and throw on an as AggregateType
it would hit. There was also no code to handle the necessary cast, then
lifted operation, then cast back.

Catch the case of nullable enums early, and turn it into a cast on the
operation on a cast to the underlying type (or int if smaller than int)

Fixes #25067