Remove Error.BindBinaryOperatorRequireTwoArguments()
BindBinaryOperation is only called by DispatchPayload, which is only
called by RuntimeBinder.BindCore, which is only called by
RuntimeBinder.Bind, which is only called with a
CSharpBinaryOperationBinder as the first argument from
CSharpBinaryOperationBinder.FallbackBinaryOperation which passes a
fixed 2-element array. Anything but 2-elements is hence impossible.

Include tests that try to pass an different number, which throw in the
DynamicMetaObjectBinder base.