Fix IndexOutOfRangeException in Socket.ConnectAsync/EndConnect (#25243)
There is a race condition in Socket.EndConnect, where it attempts to access the RemoteEndPoint of the async result before waiting for the async result to complete.  Thus if EndConnect is used immediately after BeginConnect (e.g. without a callback and just invoking it synchronosuly), we might access the remote end point before it's been initialized.