Remove some unreachable errors from Microsoft.CSharp (#25231)
* Remove Error.BindBinaryOperatorRequireTwoArguments()

BindBinaryOperation is only called by DispatchPayload, which is only
called by RuntimeBinder.BindCore, which is only called by
RuntimeBinder.Bind, which is only called with a
CSharpBinaryOperationBinder as the first argument from
CSharpBinaryOperationBinder.FallbackBinaryOperation which passes a
fixed 2-element array. Anything but 2-elements is hence impossible.

Include tests that try to pass an different number, which throw in the
DynamicMetaObjectBinder base.

* Remove Error.BindUnaryOperatorRequireOneArgument()

BindUnaryOperation is only called by DispatchPayload, which is only
called by RuntimeBinder.BindCore, which is only called by
RuntimeBinder.Bind, which is only called with a
CSharpUnaryOperationBinder as the first argument from
CSharpUnaryOperationBinder.FallbackBinaryOperation which passes a fixed
1-element array. Anything but 1 element is hence impossible.

* Remove Error.BindBinaryAssignmentRequireTwoArguments()

BindUnaryOperation is only called by
CSharpSetIndexBinder.DispatchPayload or
CSharpSetMemberBinder.DispatchPayload.

CSharpSetMemberBinder.DispatchPayload will only be calling with a fixed
2-element array. CSharpSetIndexBinder.DispatchPayload will only be
called with an array with at least 3 elements.

Anything but with fewer than 2 elements is hence impossible.

* Remove Error.BindRequireArguments()

All calls include the target in the arguments, some include more, so
there's never fewer than one arg.

* Remove check that payload to RuntimeBinder.Bind is ICSharpBinder

Call always traces back to an ICSharpBinder implementation passing this
so can never not be so.